warning: comparison between signed and unsigned integer expressions fixed
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@2857 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ cdc/modules/cdcDigitizer/src/CDCDigiModule.cc
@@ -251,7 +251,7 @@ void CDCDigiModule::event()
 
     for (int index = 0; index < cdcSimHitRel.getEntries(); index++) {
       for (int hit = 0; hit < cdcSimHitRel[index].getToIndices().size(); hit++) {
-        if (cdcSimHitRel[index].getToIndex(hit) == iterCDCMap->second->getHitNumber()) {
+        if (cdcSimHitRel[index].getToIndex(hit) == (unsigned int) iterCDCMap->second->getHitNumber()) {
           mcPartToCDCHitRel.add(cdcSimHitRel[index].getFromIndex(), iDigits);      //add entry
         }
       }
